CTHz/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
sqlmodel==0.0.22
httpx==0.27.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.5.2
python-json-logger==2.0.7
alembic==1.13.2
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.6.9
black==24.8.0
isort==5.13.2
jinja2==3.1.4
